+++ extension/frontend/src/api/installJob.ts
@@ -6,15 +6,10 @@ export type InstallResult = {
   files: string[];
 };
 
-export default async (
-  serverUuid: string,
-  jobId: string,
-  installPath: string,
-  archive = false,
-): Promise<InstallResult> => {
+export default async (serverUuid: string, jobId: string, installPath: string): Promise<InstallResult> => {
   const { data } = await axiosInstance.post(
     `/api/client/servers/${serverUuid}/calaworkshop/downloads/${jobId}/install`,
-    { install_path: installPath, archive },
+    { install_path: installPath },
   );
   return data;
 };